+++ ck2/CHANGES.md
@@ -1,5 +1,7 @@
 # Prototyping phase
 
+## Master
+
 ## V0.7.4
    - added support to detect repo, automation and artifact in a current directory 
      when using CM CLI: